feat: updata backend code

+++ src/main/java/com/example/realestatehub/dto/house/HouseQuery.java
@@ -4,12 +4,14 @@ import com.example.realestatehub.model.enums.HouseStatus;
 import lombok.Data;
 import org.springframework.data.domain.Sort;
 
+import java.math.BigDecimal;
+
 @Data
 public class HouseQuery {
     private String region;
     private HouseStatus status;
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private String layout;
     private int page = 0;
     private int size = 10;
